+++ madmom/io/midi.py
@@ -293,7 +293,7 @@ class MIDIFile(mido.MidiFile):
             tempi.insert(0, (0, DEFAULT_TEMPO))
         # tempo is given in microseconds per quarter note
         # TODO: add option to return in BPM
-        return np.asarray(tempi, np.float)
+        return np.asarray(tempi, float)
 
     @property
     def time_signatures(self):
@@ -321,7 +321,7 @@ class MIDIFile(mido.MidiFile):
             signatures.insert(0, (0, DEFAULT_TIME_SIGNATURE[0],
                                   DEFAULT_TIME_SIGNATURE[1]))
         # return time signatures
-        return np.asarray(signatures, dtype=np.float)
+        return np.asarray(signatures, dtype=float)
 
     @property
     def notes(self):
@@ -371,7 +371,7 @@ class MIDIFile(mido.MidiFile):
                 del sounding_notes[note]
 
         # sort the notes and convert to numpy array
-        return np.asarray(sorted(notes), dtype=np.float)
+        return np.asarray(sorted(notes), dtype=float)
 
     @property
     def sustain_messages(self):
